delta_data_entry/requirements.txt
flask
databricks-sdk
gunicorn